bugFixing
I just wanted to fix the add/remove button but ended up doing a TON of bug fixing. The general gist of it is that we created an object of wine to display that was identical to a wine in our list, but was not the SAME object, meaning the values of the wine would never get updated, because the object would only exist in wineDisplay. We still need to figure out how to make sure that the currentLibrary is aware of the changes globally and not just wineDisplay.